Removes ResourceSelectorHost and legacy selectors
This removes various Variable types that were tied to resource
selectors, such as GeomCache, Model, Animation, File.  Removes the
ResourceSelectorHost completely.  The two things that still appeared to
have selectors in TrackView are Audio Controls and Texture.  Fixed the
audio control selector to work via EBus and the Texture selector didn't
seem to work at all, but left it in as it was.

+++ Code/Editor/ViewportTitleDlg.cpp
@@ -449,7 +449,7 @@ void CViewportTitleDlg::AddFOVMenus(QMenu* menu, std::function<void(float)> call
 
     if (!customPresets.empty())
     {
-        for (size_t i = 0; i < customPresets.size(); ++i)
+        for (int i = 0; i < customPresets.size(); ++i)
         {
             if (customPresets[i].isEmpty())
             {
@@ -535,7 +535,7 @@ void CViewportTitleDlg::AddAspectRatioMenus(QMenu* menu, std::function<void(int,
 
     menu->addSeparator();
 
-    for (size_t i = 0; i < customPresets.size(); ++i)
+    for (int i = 0; i < customPresets.size(); ++i)
     {
         if (customPresets[i].isEmpty())
         {
@@ -684,7 +684,7 @@ void CViewportTitleDlg::AddResolutionMenus(QMenu* menu, std::function<void(int, 
 
     menu->addSeparator();
 
-    for (size_t i = 0; i < customPresets.size(); ++i)
+    for (int i = 0; i < customPresets.size(); ++i)
     {
         if (customPresets[i].isEmpty())
         {
